common data_serializer, fields file
