+++ bulbs/infographics/data_serializers.py
@@ -1,23 +1,7 @@
-from djbetty.serializers import ImageFieldSerializer
 from rest_framework import serializers
 
-from .fields import RichTextField
-
-
-class BaseEntrySerializer(serializers.Serializer):
-
-    def __init__(self, *args, **kwargs):
-        self.child_label = kwargs.pop("child_label", None)
-        super(BaseEntrySerializer, self).__init__(*args, **kwargs)
-
-
-class CopySerializer(serializers.Serializer):
-    copy = RichTextField(required=True, field_size="long")
-
-
-class EntrySerializer(BaseEntrySerializer, CopySerializer):
-    title = RichTextField(field_size="short")
-    image = ImageFieldSerializer(required=False)
+from bulbs.utils.fields import RichTextField
+from bulbs.utils.data_serializers import CopySerializer, EntrySerializer, BaseEntrySerializer
 
 
 class XYEntrySerializer(BaseEntrySerializer):
