+++ bulbs/super_features/metadata.py
@@ -9,8 +9,8 @@ from rest_framework.utils.field_mapping import ClassLookupDict
 from djbetty.serializers import ImageFieldSerializer
 
 from bulbs.content.serializers import AuthorField
+from bulbs.utils.fields import RichTextField
 from bulbs.infographics.data_serializers import CopySerializer, EntrySerializer, XYEntrySerializer
-from bulbs.infographics.fields import RichTextField
 from bulbs.super_features.serializers import BaseSuperFeatureSerializer, BaseSuperFeatureDataField
